Fix publisher and subscriber paths JC-627
