+++ apps/vanilla/src/pages/blackjack/ts/publisher.ts
@@ -1,16 +1,47 @@
+import { TSubscriber } from './publisher';
+
 /**
- * Type for subscriber entity.
- *
- * @typedef {Subscriber}
- * @template T
+ * Publisher class that manages subscribers and notifies them with updates.
  */
-export type TSubscriber<T> = {
+export class Publisher<T> {
+	private readonly subscribers: TSubscriber<T>[] = [];
 
 	/**
-	 * Method to receive a notification from the publisher.
-	 *
-	 * @param {T} message
-	 * @returns {void}
+	 * Adds a subscriber.
+	 * @param s - The subscriber to add.
 	 */
-	readonly update: (message: T) => void;
-};
+	public subscribe(s: TSubscriber<T>): void {
+		const subIndex = this.getSubscriberIndex(s);
+		if (subIndex === -1) {
+			this.subscribers.push(s);
+		}
+	}
+
+	/**
+	 * Removes a subscriber.
+	 * @param s - The subscriber to remove.
+	 */
+	public unsubscribe(s: TSubscriber<T>): void {
+		const subIndex = this.getSubscriberIndex(s);
+		if (subIndex !== -1) {
+			this.subscribers.splice(subIndex, 1);
+		}
+	}
+
+	/**
+	 * Notifies all subscribers with a message.
+	 * @param message - The message to notify subscribers with.
+	 */
+	public notify(message: T): void {
+		this.subscribers.forEach(sub => sub.update(message));
+	}
+
+	/**
+	 * Receives subscriber index.
+	 * @param s - The subscriber whose index you want to find.
+	 * @returns - The subscriber index.
+	 */
+	private getSubscriberIndex(s: TSubscriber<T>): number {
+		return this.subscribers.findIndex(sub => sub === s);
+	}
+}
